+++ uniform-geometry/tests/Uniform/Geometry_test.hs
@@ -33,11 +33,11 @@ import UniformBase
 import Uniform.GeometryFunctions
 -- import qualified Data.Geometry.Point as HP 
 
-p1 = P2d 1 (V2 1 2):: P2
+p1 = Pnv2d 1 (V2 1 2):: Pnv2
 
-test_p1 = assertEqual "P2d {_p2id = 1, _v2 = V2 1.0 2.0}" (showT p1)
+test_p1 = assertEqual "Pnv2d {_p2id = 1, _v2 = V2 1.0 2.0}" (showT p1)
 
-test_v2zero = assertEqual "P2d {_p2id = 0, _v2 = V2 0.0 0.0}" (showT (zero::P2))
+test_v2zero = assertEqual "Pnv2d {_p2id = 0, _v2 = V2 0.0 0.0}" (showT (zero::Pnv2))
 
 test_id :: IO ()
 test_id = assertEqual (1::Integer) (p1 ^. p2id)
@@ -47,17 +47,17 @@ test_v2y :: IO ()
 test_v2y = assertEqual (2.0) ((p1 ^. v2 . _y))
 
 test_toP2 :: IO ()
-test_toP2 = assertEqual p1 (tup2P2 . p2_tup_id $ p1)
+test_toP2 = assertEqual p1 (ddn_pnv2d . pnv2d_ddn $ p1)
 -- use inverse test 
 
 -- for tests five points 
 fiveV2 :: [V2d]
-fiveV2 = map p2toV2 fiveP2
+fiveV2 = map point2d_v2 fivePoint2d
 
 test_ccw_t1 :: IO ()
 test_ccw_t1 = assertBool (ccw_test (fiveV2 !! 0) (fiveV2 !! 1) (fiveV2 !! 2))
 test_ccw_t2 = assertEqual False (ccw_test (fiveV2 !! 1) (fiveV2 !! 0) (fiveV2 !! 2))
-test_ccw_t1a = assertBool (ccw_test (fiveP2 !! 0) (fiveP2 !! 1) (fiveP2 !! 2))
+test_ccw_t1a = assertBool (ccw_test (fivePoint2d !! 0) (fivePoint2d !! 1) (fivePoint2d !! 2))
 test_ccw_t3 = assertBool (ccw_test  ([0,0]::[Double]) ([5,0]::[Double]) ([10,10]::[Double]))
 
 test_scale1 = assertEqual (V2 4.0 8.0::V2d) (scale 4 (V2 1 2))